get parts at length
