+++ packages/paths/src/index.ts
@@ -1,1 +1,6 @@
 export {getLength} from './get-length';
+export {getPartAtLength} from './get-part-at-length';
+export {getParts} from './get-parts';
+export {getPointAtLength} from './get-point-at-length';
+export {getPropertiesAtLength} from './get-properties-at-length';
+export {getTangentAtLength} from './get-tangent-at-length';
